add dpo script

+++ src/start.sh
@@ -1,17 +1,40 @@
-########## Clean output director ##########
-rm -rf /data/lixubin/MiraTrainer/src/output/*
+for arg in "$@"
+do
+    echo "Script Task: $arg"
 
-########## Accelerate launch PT ##########
-# accelerate launch \
-#     --config_file /data/lixubin/MiraTrainer/src/tuner/accelerate_config.yaml \
-#     ./tuner/run_pt.py   
+    if [ $arg = "pt" ]; then
+        ########## Clean output director ##########
+        rm -rf /data/lixubin/MiraTrainer/src/output/*
+        ########## Accelerate launch PT ##########
+        accelerate launch \
+            --config_file /data/lixubin/MiraTrainer/src/tuner/accelerate_config.yaml \
+            ./tuner/run_pt.py   
+    fi
 
-########## Accelerate launch SFT ##########
-accelerate launch \
-    --config_file /data/lixubin/MiraTrainer/src/tuner/accelerate_config.yaml \
-    /data/lixubin/MiraTrainer/src/tuner/run_sft.py   
+    if [ $arg = "sft" ]; then
+        ########## Clean output director ##########
+        rm -rf /data/lixubin/MiraTrainer/src/output/*
+        ########## Accelerate launch SFT ##########
+        accelerate launch \
+            --config_file /data/lixubin/MiraTrainer/src/tuner/accelerate_config.yaml \
+            /data/lixubin/MiraTrainer/src/tuner/run_sft.py   
+    fi
 
-########## Test model with LoRA ##########
-CUDA_VISIBLE_DEVICES=1,2 \
-python /data/lixubin/MiraTrainer/src/demo/qwen_cli_demo.py \
-    -c /data/lixubin/MiraTrainer/src/output
+    if [ $arg = "dpo" ]; then
+        ########## Clean output director ##########
+        rm -rf /data/lixubin/MiraTrainer/src/output/*
+        ########## Accelerate launch DPO ##########
+        accelerate launch \
+            --config_file /data/lixubin/MiraTrainer/src/tuner/accelerate_config.yaml \
+            /data/lixubin/MiraTrainer/src/tuner/run_dpo.py
+
+    fi
+
+    if [ $arg = "cli" ]; then
+        ########## Test model with LoRA ##########
+        CUDA_VISIBLE_DEVICES=1,2 \
+        python /data/lixubin/MiraTrainer/src/demo/qwen_cli_demo.py \
+            -c /data/lixubin/MiraTrainer/src/output
+    fi
+
+done
